requirements.txt
pandas
wfdb
matplotlib
scipy
tqdm
tensorflow-gpu
seaborn
scikit-learn
